requirements.txt
pyvesync==1.4.3
PyYAML==6.0
